requirements.txt
requests
python-dotenv
pyyaml
claude-agent-sdk
